feat: add intro page section
